Lift up registry type
