+++ src/binary/ibc-setup/types.ts
@@ -1,0 +1,12 @@
+type Chain = {
+  chain_id: string;
+  prefix: string;
+  gas_price: string;
+  hd_path: string;
+  rpc: string[];
+};
+
+export type Registry = {
+  version: number;
+  chains: Record<string, Chain>;
+};
